🎨 Style: 마스터포트폴리오 타입 분리, 용도 재정의
