+++ src/services/user/user.ts
@@ -2,15 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import axiosInstance from '@/lib/axiosInstance';
 
 // 사용자 프로필 정보를 가져오는 함수
-const fetchUserProfile = async (): Promise<UserProfile> => {
+export default async function fetchUserProfile(): Promise<UserProfile> {
   const { data } = await axiosInstance.get<UserResponse>('/api/v1/users/me');
   return data.result;
-};
-
-export const useUser = () => {
-  return useQuery({
-    queryKey: ['user'],
-    queryFn: fetchUserProfile,
-    staleTime: 1000 * 60 * 5,
-  });
-};
+}
